+++ tresboncoin/data/motomag/__init__.py
@@ -1,0 +1,8 @@
+from os.path import isfile
+from os.path import dirname
+
+version_file = '{}/version.txt'.format(dirname(__file__))
+
+if isfile(version_file):
+    with open(version_file) as version_file:
+        __version__ = version_file.read().strip()
