+++ tresboncoin/data/motomag/scraping.py
@@ -1,0 +1,5 @@
+from data.motomag.scraping_to_dataframe import scraping_to_dataframe
+
+
+def scraping():
+    scraping_to_dataframe()
