+++ tresboncoin/data/motoplanete/scraping_annonces.py
@@ -4,41 +4,40 @@ from utils import *
 import requests
 from bs4 import BeautifulSoup
 import os
-import codecs
 from datetime import datetime
-import datetime
 import time
-import random
 import pandas as pd
+
+PATH_TO_CSV = os.path.dirname(os.path.abspath(__file__)).replace('/motoplante', '') + '/scraping_outputs/motoplante.csv'
+PATH_TO_LOG = os.path.dirname(os.path.abspath(__file__)).replace('/motoplante', '') + "/log.csv"
+PATH_TO_FOLDER = os.path.dirname(os.path.abspath(__file__))
+PATH_TO_IMG_FOLDER = os.path.dirname(os.path.abspath(__file__)) + '/img'
+PATH_TO_PAGES_FOLDER = os.path.dirname(os.path.abspath(__file__)) + ('/pages')
+PATH_TO_ANNONCES_FOLDER = os.path.dirname(os.path.abspath(__file__)) + ('/annonces')
+PATH_TO_INDEX = os.path.dirname(os.path.abspath(__file__)) + '/index.csv'
 
 
 def scraping_annonces():
 
-    # log file name
-    csv_name = "../log_motoplanete.csv"
-
     # init scrap
-    directory = 'pages'
+    directory = PATH_TO_PAGES_FOLDER
     #scrap = True
 
     # site to scrap
     source = 'motoplanete'
 
+    # Start time
+    start_time = datetime.now()
+
     # log update
+    log_import = pd.read_csv(PATH_TO_LOG)
     log_new = pd.DataFrame({'source': [source],
-                            'step': ['scrap pages'],
+                            'step': ['scrap annonces'],
                             'status': ['started'],
-                            'time': [datetime.datetime.now()],
+                            'time': [datetime.now()],
                             'details': [""]})
-
-    # init log
-    if os.path.isfile(csv_name) is False:
-        log_new.to_csv(csv_name, index=False)
-    else:
-        log_import = pd.read_csv(csv_name)
-        log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
-
+    log = log_import.append(log_new, ignore_index=True)
+    log.to_csv(PATH_TO_LOG, index=False)
 
     # load csv if exists or starting from template
     data_motoplanete = "../motoplanete.csv"
@@ -85,30 +84,28 @@ def scraping_annonces():
             os.remove(directory+"/"+html_file)
 
         # End time
-        end_time = datetime.datetime.now()
+        end_time = datetime.now()
         td = end_time - start_time
 
         # log update
-        log_import = pd.read_csv(csv_name)
+        log_import = pd.read_csv(PATH_TO_LOG)
         log_new = pd.DataFrame({'source': [source],
-                                'step': ['scrap annonces'],
+                                'step': ['scrap pages'],
                                 'status': ['completed'],
-                                'time': [datetime.datetime.now()],
-                                'details': [f"{td.seconds/60} minutes elapsed, {count} annonces scrapped"]})
+                                'time': [datetime.now()],
+                                'details': [f"{td.seconds/60} minutes elapsed"]})
         log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
+        log.to_csv(PATH_TO_LOG, index=False)
 
     except (ValueError, TypeError, NameError, KeyError, RuntimeWarning) as err:
         # log update
-        log_import = pd.read_csv(csv_name)
+        log_import = pd.read_csv(PATH_TO_LOG)
         log_new = pd.DataFrame({'source': [source],
-                                'step': ['scrap annonces'],
-                                'status': ['error'],
-                                'time': [datetime.datetime.now()],
+                               'status': ['error'],
+                                'time': [datetime.now()],
                                 'details': [err]})
         log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
-    return
+        log.to_csv(PATH_TO_LOG, index=False)
 
 
 if __name__ == "__main__":
