+++ tresboncoin/data/motoplanete/scraping_to_dataframe.py
@@ -1,54 +1,45 @@
-import requests
 from parameters import *
 from scraper import *
 from bs4 import BeautifulSoup
-import requests
-from PIL import Image
 import os
-import glob
 import pandas as pd
-import datetime
-import time
-import random
-import codecs
-import shutil
 from datetime import datetime
 import subprocess
 
 
+PATH_TO_CSV = os.path.dirname(os.path.abspath(__file__)).replace('/motoplante', '') + '/scraping_outputs/motoplante.csv'
+PATH_TO_LOG = os.path.dirname(os.path.abspath(__file__)).replace('/motoplante', '') + "/log.csv"
+PATH_TO_FOLDER = os.path.dirname(os.path.abspath(__file__))
+PATH_TO_IMG_FOLDER = os.path.dirname(os.path.abspath(__file__)) + '/img'
+PATH_TO_PAGES_FOLDER = os.path.dirname(os.path.abspath(__file__)) + ('/pages')
+PATH_TO_ANNONCES_FOLDER = os.path.dirname(os.path.abspath(__file__)) + ('/annonces')
+PATH_TO_INDEX = os.path.dirname(os.path.abspath(__file__)) + '/index.csv'
+
+
 def scraping_to_dataframe():
 
-
-    # log file name
-    csv_name = "../log_motoplanete.csv"
-
-
     # directory of html annonces
-    directory = 'annonces'
-
+    directory = PATH_TO_ANNONCES_FOLDER
 
     # site to scrap
     source = 'motoplanete'
 
+    # Start time
+    start_time = datetime.now()
 
     # log update
+    log_import = pd.read_csv(PATH_TO_LOG)
     log_new = pd.DataFrame({'source': [source],
-                            'step': ['scrap pages'],
+                            'step': ['scrap annonces'],
                             'status': ['started'],
                             'time': [datetime.now()],
                             'details': [""]})
-
-    # init log
-    if os.path.isfile(csv_name) is False:
-        log_new.to_csv(csv_name, index=False)
-    else:
-        log_import = pd.read_csv(csv_name)
-        log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
+    log = log_import.append(log_new, ignore_index=True)
+    log.to_csv(PATH_TO_LOG, index=False)
 
 
     # load csv if exists or starting from template
-    data_motoplanete = "../motoplanete.csv"
+    data_motoplanete = PATH_TO_CSV
     #
     if os.path.isfile(data_motoplanete) is False:
         df = motoplanete_announce_template.copy()
@@ -56,11 +47,9 @@ def scraping_to_dataframe():
         data = pd.read_csv(data_motoplanete)
         df = motoplanete_announce_template.copy()
 
-
     try:
         # Start time of the pages scrapping
         start_time = datetime.now()
-
 
         for html_file in [file for file in os.listdir(directory) if file.endswith(".html")]:
 
@@ -73,8 +62,6 @@ def scraping_to_dataframe():
 
             # get soup
             soup_ = BeautifulSoup(readable_html, 'html.parser')
-
-
 
             # fill df with defined functions
             announce_scrap = motoplanete_page_scraper(announce=soup_)
@@ -103,39 +90,36 @@ def scraping_to_dataframe():
             # concatenate to csv and write
             try:
                 data = pd.concat([data, df], axis=0)
-                data.to_csv(path_or_buf = data_motoplanete, index=False)
+                data.to_csv(path_or_buf=data_motoplanete, index=False)
             except:
-                df.to_csv(path_or_buf = data_motoplanete, index=False)
+                df.to_csv(path_or_buf=data_motoplanete, index=False)
 
             # deplacer le fichier html traité dans le vault
             subprocess.run(["mv", os.path.join(directory, html_file), "vault"])
-
 
         # End time
         end_time = datetime.now()
         td = end_time - start_time
 
         # log update
-        log_import = pd.read_csv(csv_name)
+        log_import = pd.read_csv(PATH_TO_LOG)
         log_new = pd.DataFrame({'source': [source],
-                                'step': ['to dataframe'],
+                                'step': ['scrap pages'],
                                 'status': ['completed'],
                                 'time': [datetime.now()],
                                 'details': [f"{td.seconds/60} minutes elapsed"]})
         log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
-
+        log.to_csv(PATH_TO_LOG, index=False)
 
     except (ValueError, TypeError, NameError, KeyError, RuntimeWarning) as err:
         # log update
-        log_import = pd.read_csv(csv_name)
+        log_import = pd.read_csv(PATH_TO_LOG)
         log_new = pd.DataFrame({'source': [source],
-                                'step': ['to dataframe'],
-                                'status': ['error'],
+                               'status': ['error'],
                                 'time': [datetime.now()],
                                 'details': [err]})
         log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
+        log.to_csv(PATH_TO_LOG, index=False)
 
 
 if __name__ == "__main__":
