+++ tresboncoin/data/paruvendu/scraping_to_dataframe.py
@@ -14,40 +14,37 @@ import shutil
 from datetime import datetime
 import subprocess
 
+PATH_TO_CSV = os.path.dirname(os.path.abspath(__file__)).replace('/paruvendu', '') + '/scraping_outputs/paruvendu.csv'
+PATH_TO_LOG = os.path.dirname(os.path.abspath(__file__)).replace('/paruvendu', '') + "/log.csv"
+PATH_TO_FOLDER = os.path.dirname(os.path.abspath(__file__))
+PATH_TO_IMG_FOLDER = os.path.dirname(os.path.abspath(__file__)) + '/img'
+PATH_TO_PAGES_FOLDER = os.path.dirname(os.path.abspath(__file__)) + ('/pages')
+PATH_TO_ANNONCES_FOLDER = os.path.dirname(os.path.abspath(__file__)) + ('/annonces')
+PATH_TO_INDEX = os.path.dirname(os.path.abspath(__file__)) + '/index.csv'
+
 
 def scraping_to_dataframe():
-
-
-    # log file name
-    csv_name = "../log_pv.csv"
-
-
-    # directory of html annonces
-    directory = 'annonces'
-
 
     # site to scrap
     source = 'paruvendu'
 
+    directory =
+
+    # Start time
+    start_time = datetime.now()
 
     # log update
+    log_import = pd.read_csv(PATH_TO_LOG)
     log_new = pd.DataFrame({'source': [source],
-                            'step': ['scrap pages'],
+                            'step': ['scrap annonces'],
                             'status': ['started'],
                             'time': [datetime.now()],
                             'details': [""]})
-
-    # init log
-    if os.path.isfile(csv_name) is False:
-        log_new.to_csv(csv_name, index=False)
-    else:
-        log_import = pd.read_csv(csv_name)
-        log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
-
+    log = log_import.append(log_new, ignore_index=True)
+    log.to_csv(PATH_TO_LOG, index=False)
 
     # load csv if exists or starting from template
-    data_paruvendu = "../paruvendu.csv"
+    data_paruvendu = PATH_TO_CSV
     #
     if os.path.isfile(data_paruvendu) is False:
         df = paruvendu_announce_template.copy()
@@ -55,11 +52,9 @@ def scraping_to_dataframe():
         data = pd.read_csv(data_paruvendu)
         df = paruvendu_announce_template.copy()
 
-
     try:
         # Start time of the pages scrapping
         start_time = datetime.now()
-
 
         for html_file in [file for file in os.listdir(directory) if file.endswith(".html")]:
 
@@ -72,8 +67,6 @@ def scraping_to_dataframe():
 
             # get soup
             soup_ = BeautifulSoup(readable_html, 'html.parser')
-
-
 
             # fill df with defined functions
             announce_scrap = paruvendu_page_scraper(announce=soup_)
@@ -112,32 +105,29 @@ def scraping_to_dataframe():
             # deplacer le fichier html traité dans le vault
             subprocess.run(["mv", os.path.join(directory, html_file), "vault"])
 
-
         # End time
         end_time = datetime.now()
         td = end_time - start_time
 
         # log update
-        log_import = pd.read_csv(csv_name)
+        log_import = pd.read_csv(PATH_TO_LOG)
         log_new = pd.DataFrame({'source': [source],
-                                'step': ['to dataframe'],
+                                'step': ['scrap pages'],
                                 'status': ['completed'],
                                 'time': [datetime.now()],
                                 'details': [f"{td.seconds/60} minutes elapsed"]})
         log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
-
+        log.to_csv(PATH_TO_LOG, index=False)
 
     except (ValueError, TypeError, NameError, KeyError, RuntimeWarning) as err:
         # log update
-        log_import = pd.read_csv(csv_name)
+        log_import = pd.read_csv(PATH_TO_LOG)
         log_new = pd.DataFrame({'source': [source],
-                                'step': ['to dataframe'],
-                                'status': ['error'],
+                               'status': ['error'],
                                 'time': [datetime.now()],
                                 'details': [err]})
         log = log_import.append(log_new, ignore_index=True)
-        log.to_csv(csv_name, index=False)
+        log.to_csv(PATH_TO_LOG, index=False)
 
 
 if __name__ == "__main__":
